[community-package] Address review: fold generate into resourcedocsgen, check version once
Move provider_name into models so resourcedocsgen owns both the single-run
and the batch-publish path; delete generate.py and call resourcedocsgen.publish
from cli.

Validate the release tag against SAFE_VERSION at the top of probe_installs, so
the version derived for the npm/pip/go probes is checked rather than only the
plugin path.

## Test plan

- `python3 -m unittest discover -p 'test_*.py'` (12 tests)
- `mypy --strict *.py`
